+++ src/zabbix-fetch.ts
@@ -2,12 +2,18 @@ export type ZabbixRequestParams = {
   [propName: string]: any
 }
 
+export type ZabbixRequestOptions = {
+  authToken?: string
+  timeout?: number
+}
+
 function zabbixFetch(
   url: string,
   apiMethod: string,
   params: ZabbixRequestParams = {},
-  authToken?: string
+  options?: ZabbixRequestOptions
 ): Promise<Response> {
+  const { authToken, timeout } = options || {}
   const data = {
     jsonrpc: '2.0',
     method: apiMethod,
@@ -15,12 +21,14 @@ function zabbixFetch(
     auth: authToken || null,
     id: Math.floor(Math.random() * 100)
   }
+  const signal = timeout ? AbortSignal.timeout(timeout) : undefined
   return fetch(url, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json-rpc'
     },
-    body: JSON.stringify(data)
+    body: JSON.stringify(data),
+    signal
   })
 }