+++ src/zabbix.ts
@@ -1,4 +1,7 @@
-import zabbixFetch, { ZabbixRequestParams } from './zabbix-fetch'
+import zabbixFetch, {
+  ZabbixRequestOptions,
+  ZabbixRequestParams
+} from './zabbix-fetch'
 import { ZabbixResponse } from './types/zabbix-response'
 import { IZabbix } from './interfaces/Zabbix.interface'
 import { ZabbixUserLoginParams } from './types/user'
@@ -7,13 +10,14 @@ export default class Zabbix implements IZabbix {
   private readonly username: string
   private readonly password: string
   protected readonly apiURL: string
+  readonly options: ZabbixRequestOptions
   authToken: string = ''
 
   constructor(
     url: string,
     username: string,
     password: string,
-    authToken?: string
+    options?: ZabbixRequestOptions
   ) {
     if (!url) {
       throw new Error('url parameter not found in constructor')
@@ -27,6 +31,8 @@ export default class Zabbix implements IZabbix {
     this.apiURL = url
     this.username = username
     this.password = password
+    this.options = options || {}
+    const { authToken } = options || {}
     if (authToken) {
       this.authToken = authToken
     }
@@ -38,7 +44,9 @@ export default class Zabbix implements IZabbix {
       password: params.password
     })
     if (!response.ok) {
-      throw new Error(`Login HTTP error! status: ${response.status}`)
+      throw new Error(
+        `Login HTTP error! [${response.status}]: ${response.statusText}`
+      )
     }
     const res = (await response.json()) as ZabbixResponse<string>
     return res
@@ -53,19 +61,30 @@ export default class Zabbix implements IZabbix {
     params: ZabbixRequestParams
   ): Promise<ZabbixResponse<any>> {
     const response = await this.getDataAPI(apiMethod, params)
-    if (!response.ok) {
-      throw new Error(`${response.status}: ${response.statusText}`)
-    }
     const data: ZabbixResponse<any> =
       (await response.json()) as ZabbixResponse<any>
     return data
   }
 
   async getDataAPI(apiMethod: string, params: ZabbixRequestParams) {
-    if (!this.authToken) {
-      await this.getAuthToken()
+    try {
+      if (!this.authToken) {
+        await this.getAuthToken()
+      }
+      const response = await zabbixFetch(this.apiURL, apiMethod, params, {
+        ...this.options,
+        authToken: this.authToken
+      })
+      if (!response.ok) {
+        throw new Error(`[${response.status}]: ${response.statusText}`)
+      }
+      return response
+    } catch (err) {
+      if (err instanceof Response) {
+        throw new Error(`[${err.status}]: ${err.statusText}`)
+      }
+      throw err
     }
-    return zabbixFetch(this.apiURL, apiMethod, params, this.authToken)
   }
 
   private async getAuthToken(): Promise<void> {